+++ core/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for `TransactionData::Blob`
+- Add `error::TxPreconditionError`
+- Add `transfer::deploy_check`
+- Add `transfer::blob_check`
 
 ## [1.3.0] - 2025-04-17